PDF of the PCB 001
This is a single sided PCB about 80x45mm and fits a standard ebay extruded aluminum enclosure. Photos of an assembled until will be available "real soon now".   [1]
The PCB is a 2 layer board. The relay is thought to work through 1296 and possibly 2304.  TBD!

This project released under terms of the TAPR NCL. A copy is available on this site.

[1] Who is the famous write who used this phrase nearly every month?  <-- my attempt at humor